rename kungfubase to kungfu/base
