+++ srcs/go/kungfu/base/dtype.c
