+++ srcs/go/kungfu/base/kungfu
@@ -1,0 +1,1 @@
+../../../cpp/include/kungfu
