+++ srcs/go/kungfu/base/op.cpp
