+++ srcs/go/kungfu/base/op.go
@@ -1,4 +1,4 @@
-package kungfubase
+package base
 
 import "unsafe"
