+++ srcs/go/kungfu/session/session.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 
 	"github.com/lsds/KungFu/srcs/go/kungfu/execution"
-	kb "github.com/lsds/KungFu/srcs/go/kungfubase"
+	kb "github.com/lsds/KungFu/srcs/go/kungfu/base"
 	"github.com/lsds/KungFu/srcs/go/log"
 	"github.com/lsds/KungFu/srcs/go/plan"
 	rch "github.com/lsds/KungFu/srcs/go/rchannel"
