+++ srcs/go/kungfu/session/strategy.go
@@ -1,7 +1,7 @@
 package session
 
 import (
-	kb "github.com/lsds/KungFu/srcs/go/kungfubase"
+	kb "github.com/lsds/KungFu/srcs/go/kungfu/base"
 	"github.com/lsds/KungFu/srcs/go/plan"
 )
