--- srcs/go/kungfubase/kungfu
@@ -1,1 +1,0 @@
-../../cpp/include/kungfu
