+++ srcs/go/plan/id.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"strconv"
 
-	kb "github.com/lsds/KungFu/srcs/go/kungfubase"
+	kb "github.com/lsds/KungFu/srcs/go/kungfu/base"
 )
 
 // PeerID is the unique identifier of a peer.
