+++ tests/go/cmd/kungfu-test-public-apis/kungfu-test-public-apis.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/lsds/KungFu/srcs/go/kungfu/peer"
 	"github.com/lsds/KungFu/srcs/go/kungfu/session"
-	kb "github.com/lsds/KungFu/srcs/go/kungfubase"
+	kb "github.com/lsds/KungFu/srcs/go/kungfu/base"
 	"github.com/lsds/KungFu/srcs/go/utils"
 )
